+++ src/runtime/getTarget.js
@@ -1,0 +1,29 @@
+const memo = {};
+
+/* istanbul ignore next  */
+function getTarget(target) {
+  if (typeof memo[target] === "undefined") {
+    let styleTarget = document.querySelector(target);
+
+    // Special case to return head of iframe instead of iframe itself
+    if (
+      window.HTMLIFrameElement &&
+      styleTarget instanceof window.HTMLIFrameElement
+    ) {
+      try {
+        // This will throw an exception if access to iframe is blocked
+        // due to cross-origin restrictions
+        styleTarget = styleTarget.contentDocument.head;
+      } catch (e) {
+        // istanbul ignore next
+        styleTarget = null;
+      }
+    }
+
+    memo[target] = styleTarget;
+  }
+
+  return memo[target];
+}
+
+module.exports = getTarget;
