+++ src/runtime/singletonStyleApi.js
@@ -1,0 +1,61 @@
+/* istanbul ignore next  */
+const replaceText = (function replaceText() {
+  const textStore = [];
+
+  return function replace(index, replacement) {
+    textStore[index] = replacement;
+
+    return textStore.filter(Boolean).join("\n");
+  };
+})();
+
+/* istanbul ignore next  */
+function apply(style, index, remove, obj) {
+  const css = remove
+    ? ""
+    : obj.media
+    ? `@media ${obj.media} {${obj.css}}`
+    : obj.css;
+
+  // For old IE
+  /* istanbul ignore if  */
+  if (style.styleSheet) {
+    style.styleSheet.cssText = replaceText(index, css);
+  } else {
+    const cssNode = document.createTextNode(css);
+    const childNodes = style.childNodes;
+
+    if (childNodes[index]) {
+      style.removeChild(childNodes[index]);
+    }
+
+    if (childNodes.length) {
+      style.insertBefore(cssNode, childNodes[index]);
+    } else {
+      style.appendChild(cssNode);
+    }
+  }
+}
+
+const singletonData = {
+  singleton: null,
+  singletonCounter: 0,
+};
+
+/* istanbul ignore next  */
+function domApi(options) {
+  // eslint-disable-next-line no-undef,no-use-before-define
+  const styleIndex = singletonData.singletonCounter++;
+  const style =
+    // eslint-disable-next-line no-undef,no-use-before-define
+    singletonData.singleton ||
+    // eslint-disable-next-line no-undef,no-use-before-define
+    (singletonData.singleton = options.insertStyleElement(options));
+
+  return {
+    update: apply.bind(null, style, styleIndex, false),
+    remove: apply.bind(null, style, styleIndex, true),
+  };
+}
+
+module.exports = domApi;
